# swagger.yaml
swagger: "2.0"

info:
  version: 1.0.0
  title: ABONNEMENTS API
  description: L'API permettant de gerer les abonnements clients

schemes:
  - https
host: localhost:3332
basePath: /

paths:
  /employe:
    post:
      summary: Créer un nouvel employé.
      tags:
        - Employe
      parameters:
        - in: body
          name: employeInformations
          required: true
          schema:
            $ref: '#/definitions/employeCreation'

      responses:
        200:
          description: Employé créé
        500:
          description: Erreur, voir descriptif de l'erreur

  /employe/authenticate:
    post:
      summary: Authentifie un employé.
      tags:
        - Employe
      parameters:
        - in: body
          name: employeAuthentifications
          required: true
          schema:
            $ref: '#/definitions/employeAuthentification'

      responses:
        200:
          description: Employé authentifié
          schema:
            $ref: '#/definitions/employe'
        500:
          description: Erreur, voir descriptif de l'erreur

  /client:
    post:
      summary: Créer un nouveau client.
      tags:
        - Client
      parameters:
        - in: body
          name: clientInformations
          required: true
          schema:
            $ref: '#/definitions/clientCreation'

      responses:
        200:
          description: Client créé
        500:
          description: Erreur, voir descriptif de l'erreur

  /client/authenticate:
    post:
      summary: Authentifie un client.
      tags:
        - Client
      parameters:
        - in: body
          name: clientAuthentification
          required: true
          schema:
            $ref: '#/definitions/client'

      responses:
        200:
          description: Employé authentifié
          schema:
            $ref: '#/definitions/employe'
        500:
          description: Erreur, voir descriptif de l'erreur



definitions:
  employeCreation:
    type: object
    properties:
      nom:
        type: string
      prenom:
        type: string
      login:
        type: string
      password:
        type: string

  employeAuthentification:
    type: object
    properties:
      login:
        type: string
      password:
        type: string

  employe:
    type: object
    properties:
      nom:
        type: string
      prenom:
        type: string
      login:
        type: string
      role:
        type: string
      token:
        type: string

  clientCreation:
    type: object
    properties:
      nom:
        type: string
      prenom:
        type: string
      displayName:
        type: string
      login:
        type: string
        format: email
      dateNaissance:
        type: string
        pattern: /([0-9]{2})-(?:[0-9]{2})-([0-9]{4})/
        example: "14-01-1993"
      lieuNaissance:
        type: string
      rue:
        type: string
      cp:
        type: string
      ville:
        type: string
      password:
        type: string

  client:
    type: object
    properties:
      id:
        type: integer
      nom:
        type: string
      prenom:
        type: string
      displayName:
        type: string
      login:
        type: string
        format: email
      dateNaissance:
        type: string
        pattern: /([0-9]{2})-(?:[0-9]{2})-([0-9]{4})/
        example: "14-01-1993"
      lieuNaissance:
        type: string
      rue:
        type: string
      cp:
        type: string
      ville:
        type: string
      password:
        type: string
      role:
        type: string

  clientAuthentification:
    type: object
    properties:
      login:
        type: string
      password:
        type: string